+++ src/main/java/org/kesler/simplereg/dao/impl/GenericDAOImpl.java
@@ -78,7 +78,7 @@ public class GenericDAOImpl<T extends AbstractEntity> implements GenericDAO <T> 
 
 	}
 
-	public T getItemById(int id) {
+	public T getItemById(long id) {
 		T item = null;
 
 		notifyListeners(DAOState.CONNECTING);
